Creación de Informe en cada una de las capas

+++ ProgramacionTP_CS_API_Mongo/Models/Informe.cs
@@ -5,13 +5,13 @@ namespace ProgramacionTP_CS_API_Mongo.Models
     public class Informe
     {
         [JsonPropertyName("horarios")]
-        public int Horarios { get; set; } = 0;
+        public long Horarios { get; set; } = 0;
 
         [JsonPropertyName("autobuses")]
-        public int Autobuses { get; set; } =0;
+        public long Autobuses { get; set; } =0;
 
         [JsonPropertyName("cargadores")]
-        public int Cargadores { get; set; } = 0;
+        public long Cargadores { get; set; } = 0;
 
         [JsonPropertyName("operacion_autobuses")]
         public float Operacion_autobuses { get; set; } = 0;
